Display previews of edited content inside device frames
Summary: Previously, previews would be displayed in boxes of different widths for different devices. Now, previews are placed inside of device frames that both set the dimensions for the content inside and make it easier to know what type of device is being previewed. Also adds the device option of "No Frame", for having the previous experience of no device frame.

Test Plan:
- Verified in Perseus that when editing questions and hints, the rendered preview is displayed in a device frame