Delay mounting preview iframes until the page has loaded so that the location.href is correct
Summary: Previously, iframes rendered in the initial load would fail to initialize because the location would not be the same as the parent. Now, because they are mounted after the parent has loaded, the location is correct and the iframes are able to load.

Test Plan: - Verified in webapp that iframes load with the same location as their parent

Reviewers: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28787