Implemented widgets
Summary:
This commit is large, and the intended review process involves simply expanding
the entire `src/perseus-editor.jsx` file and reading through it.  Comments have
been placed to help it be understandable from that viewpoint.  There is
a lot of Draft-js specific implementation details, therefore it might
help in understanding things to see the big picture (the big file).

The reason it is large is because the experimentation that I did initially ended
up having to be scrapped.  I was initially storing the widget metadata within
entities themselves, however the issue arrives with copy-paste, because while
copy-paste of entities is meant to work in Draft.js, it did not appear to work
when copying from one editor to another editor.

To deal with this, widgets are instead stored as state within `PerseusEditor`,
and entities only store widget IDs.  Copy/paste will be done in a similar
fashion to how it was done before, using localstorage.

One improvement thats already come for free is that undo/redo is working very
well in all cases.  Users can delete text, undo, and the widgets will still
exist.

Test Plan:
Open a blank perseus page
Insert a widget
Add custom properties to it
Insert another widget of the same type, observe the id being unique
Add some custom proeprties to it
Backspace from the edge of a widget, observe the entire widget be removed
Press Cmd-Z, which brings the widget back
Select both widgets, delete, then Cmd-Z again to see it all come back again
Delete a widget by clicking the trash bin icon on the widget editor
Observe it being deleted, then Cmd-Z and observe it returning

Type in normal paragraphs
Undo, observe it not being letter by letter anymore, but more in chunks, which
is feels lot nicer to use.

Reviewers: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D29788