Moved the parent updating into a throttled function
Summary:
Since updating the content and widget lists involves iterating through
every single content block, this resulted in performance issues.  To fix this,
I introduced a small delay between when the content is updated and when the
display shows it.

If content creators rely on the preview updating live, then I can disable this,
but we'll find out during the test

Test Plan:
Exponentially copy/paste some text that has 2 widgets until it is 64 widgets
Try typing
Observe it not acting impossibly slow, and updating only after the user has stopped typing

Reviewers: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D29842