Made fixes mentioned in previous diffs
Summary:
for the `_insertNewWidget`, I decided to have `draftDataParams`, and return
both the new editorState and the new contentState.

I might do this for all the other functions to remove the need for
EditorState.push.  What this allows us to do is have a much more robust set of
helper functions.

The idea is as follows:

The main data about the editor that matters is the `editorState`, the
`contentState`, and the current `selection`.

The issue with returning an `editorState` object is that you can no longer
chain helper functions together, since each `EditorState.push` results in
a new `undo` state added, and we wouldnt' want `undo`s to bring you to an
intermediate state.

The issue with returning a `contentState` object is that you now have to
`EditorState.push` outside of the function, which can become a pain

Another issue is that in certain cases we want the function to use the
current `editorState`/`contentState`/`selection`, while in other cases
we need to pass in a custom one in as the helper function is being used
in the middle of other operations.

The solution used here is therefore:
Consider the core objects: `{editorState, contentState, selection}` to be "draftData"
Take in a `draftData` parameter that allows overriding of any of those aspects
Any that aren't specified default to the current values of those objects (using
this.state)

Return an entire `draftData`, with both the `editorState` and the
`contentState`.  This allows for both the avoiding of `EditorState.push` for
trivial operations, as well as being able to use helper functions in
intermediate steps, since they can just ignore the `editorState` return and use
the `contentState`

Test Plan:
Basic typing, copy/paste of 0-2 widgets, check tab completion with
dropdown

Reviewers: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D30008