Add new grafting area to target with skill/subscore box
Summary: We are adding a new feature to SAT review-mode that surfaces the skill and subscore category associated with each question in SAT diagnostics, timed minisections, skill practice and full exams. To appropriately place this information in the question UI, we need a new grafting area underneath the two column question header.

Test Plan: In isolation, this div does nothing. Will test in coordination with the diff in webapp that actually grafts the new skillSubscoreBox component onto this grafting area.

Reviewers: emily, mattdunnrankin

Subscribers: lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D34385