Replaced static CSS class with a dynamic name based on mouse position
Summary:
The original code specified a cursor CSS class name that didn't
correctly indicate the behaviour; it showed the move cursor along
the circle's perimeter, when what really happens is a resizing.
This new approach uses the mousemove event along with the mouse
position to dynamically set the class using a relative position.

Test Plan:
- make build
- Verified in a browser that the cursor changes correctly along
the edge of the circle
- Moved the circle around and verified that the behaviour is still
correct

Reviewers: emily, aasmund

Reviewed By: emily, aasmund

Subscribers: aasmund, jared

Differential Revision: https://phabricator.khanacademy.org/D36331